% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{evolution_prix}
\alias{evolution_prix}
\title{Évolution du meilleur prix par option et par jour de collecte}
\usage{
evolution_prix(offres)
}
\arguments{
\item{offres}{Table \code{offres}.}
}
\value{
Un data.frame \code{date_collecte}, \code{option}, \code{prix_min}.
}
\description{
Évolution du meilleur prix par option et par jour de collecte
}
