% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{format_duree}
\alias{format_duree}
\title{Formater une durée en minutes en « 14 h 05 »}
\usage{
format_duree(minutes)
}
\arguments{
\item{minutes}{Durée en minutes.}
}
\description{
Formater une durée en minutes en « 14 h 05 »
}
