% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{combinaisons}
\alias{combinaisons}
\title{Toutes les combinaisons de recherche décrites par la configuration}
\usage{
combinaisons(recherches)
}
\arguments{
\item{recherches}{L'élément \code{recherches} de \code{\link[=lire_config]{lire_config()}}.}
}
\value{
Un data.frame, une ligne par recherche possible.
}
\description{
Toutes les combinaisons de recherche décrites par la configuration
}
